# api/openapi.yaml
openapi: 3.1.0
info:
  title: Nabd HMS API
  version: "1.0"
  description: >
    Foundation-layer contract (Auth, RBAC/Staff, Tenant/Branch, Patients, Audit).
    See README.md for conventions. Deny-by-default authorization: every
    operation below declares x-required-permission; the gateway rejects
    anything not explicitly granted.
servers:
  - url: https://api.nabd.health/v1
    description: production
  - url: https://api.staging.nabd.health/v1
    description: staging
security:
  - bearerAuth: []
...
paths:
...
  /staff/{id}:
    get:
      tags: [Staff]
      x-required-permission: staff:view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: "Staff", content: { application/json: { schema: { $ref: "#/components/schemas/Staff" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Staff]
      summary: Update role, scope (own-patients-only vs. all), or field-level grants
      x-required-permission: staff:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/StaffPatch" } } }
      responses:
        "200": { description: "Staff", content: { application/json: { schema: { $ref: "#/components/schemas/Staff" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /roles:
    get:
      tags: [Roles]
      summary: List built-in and custom roles for the tenant
      x-required-permission: staff:view
      responses:
        "200": { description: "List of Role", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Role" } } } } }
    post:
      tags: [Roles]
      summary: Create a custom role from the permission matrix (module x action)
      x-required-permission: staff:create
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RoleWrite" } } }
      responses:
        "201": { description: "Role", content: { application/json: { schema: { $ref: "#/components/schemas/Role" } } } }
  /roles/{id}:
    get:
      tags: [Roles]
      x-required-permission: staff:view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: "Role", content: { application/json: { schema: { $ref: "#/components/schemas/Role" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Roles]
      x-required-permission: staff:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/RoleWrite" } } }
      responses:
        "200": { description: "Role", content: { application/json: { schema: { $ref: "#/components/schemas/Role" } } } }
        "400":
          description: Attempted edit of a built-in role, or a grant beyond the editor's own permissions (privilege escalation blocked)
          content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }

  # ------------------------------------------------------------ Tenants
...
  /doctors/{id}/working-hours:
    get:
      tags: [Queue]
      summary: Weekly recurring availability template for a doctor
      x-required-permission: queue:view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: "Working hours", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/WorkingHours" } } } } }
    post:
      tags: [Queue]
      x-required-permission: queue:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/WorkingHoursWrite" } } }
      responses:
        "201": { description: "Working hours block created", content: { application/json: { schema: { $ref: "#/components/schemas/WorkingHours" } } } }
  /doctors/{id}/leave:
    get:
      tags: [Queue]
      summary: Blocked date ranges for a doctor (leave)
      x-required-permission: queue:view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: "Leave ranges", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/DoctorLeave" } } } } }
    post:
      tags: [Queue]
      summary: "Block a date range; existing bookings in that range are not auto-cancelled (NB-074 — reschedule flagging deferred)"
      x-required-permission: queue:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/DoctorLeaveWrite" } } }
      responses:
        "201": { description: "Leave recorded", content: { application/json: { schema: { $ref: "#/components/schemas/DoctorLeave" } } } }
  /doctors/{id}/availability:
    get:
      tags: [Queue]
      summary: Bookable slots for a doctor on a given date (working hours minus leave minus already-booked slots)
      x-required-permission: queue:view
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Available slot start times
          content: { application/json: { schema: { type: array, items: { type: string, format: date-time } } } }
  /appointments:
    get:
      tags: [Queue]
      x-required-permission: queue:view
      parameters:
        - name: doctorId
          in: query
          schema: { type: string, format: uuid }
        - name: patientId
          in: query
          schema: { type: string, format: uuid }
        - name: date
          in: query
          schema: { type: string, format: date }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Page of appointments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Appointment" } }
                  page: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [Queue]
      summary: "Book a slot — concurrency-safe: a DB constraint, not an app-level check, prevents double-booking (NB-067)"
      x-required-permission: queue:create
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/AppointmentWrite" } } }
      responses:
        "201": { description: "Booked", content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
        "409": { description: "Slot no longer available", content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
  /appointments/{id}:
    get:
      tags: [Queue]
      x-required-permission: queue:view
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      responses:
        "200": { description: "Appointment", content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /appointments/{id}/cancel:
    post:
      tags: [Queue]
      summary: "Reason mandatory; slot released immediately (NB-072)"
      x-required-permission: queue:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 3 } } }
      responses:
        "200": { description: "Cancelled", content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /appointments/{id}/reschedule:
    post:
      tags: [Queue]
      summary: Cancels the original and books a new slot atomically
      x-required-permission: queue:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [newStartTime], properties: { newStartTime: { type: string, format: date-time } } }
      responses:
        "200": { description: "New appointment", content: { application/json: { schema: { $ref: "#/components/schemas/Appointment" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { description: "New slot no longer available", content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
  /queue/check-in:
    post:
      tags: [Queue]
      summary: "Issues a token — scheduled (appointmentId given) or walk-in (patientId + doctorId only) (NB-068)"
      x-required-permission: queue:create
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CheckInRequest" } } }
      responses:
        "201": { description: "Queue entry created", content: { application/json: { schema: { $ref: "#/components/schemas/QueueEntry" } } } }
  /queue:
    get:
      tags: [Queue]
      summary: Today's queue (Arrivals screen)
      x-required-permission: queue:view
      parameters:
        - name: doctorId
          in: query
          schema: { type: string, format: uuid }
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        "200": { description: "Queue entries", content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/QueueEntry" } } } } }
  /queue/{id}/status:
    patch:
      tags: [Queue]
      summary: "State machine transition — illegal transitions rejected, every transition audited (NB-069)"
      x-required-permission: queue:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [waiting, vitals_pending, vitals_done, in_consult, checkout_pending, completed, no_show] } }
      responses:
        "200": { description: "Updated", content: { application/json: { schema: { $ref: "#/components/schemas/QueueEntry" } } } }
        "400": { description: "Illegal transition", content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /queue/{id}/reorder:
    post:
      tags: [Queue]
      summary: "Priority flag with mandatory reason; reorders the queue (NB-068/NB-092)"
      x-required-permission: queue:edit
      parameters: [{ $ref: "#/components/parameters/PathId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [priority, reason]
              properties: { priority: { type: boolean }, reason: { type: string, minLength: 3 } }
      responses:
        "200": { description: "Updated", content: { application/json: { schema: { $ref: "#/components/schemas/QueueEntry" } } } }
        "404": { $ref: "#/components/responses/NotFound" }

  # -------------------------------------------------------------- Audit
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token. Carries tenantId, staffId, roleId and permission set as
        claims — the server trusts only these, never client-supplied
        tenant/role headers.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    NotFound:
      description: Not found (or out of scope — see per-endpoint note)
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
...
  schemas:
    Problem:
      type: object
      description: RFC 7807
      properties:
        type: { type: string, format: uri, example: "https://api.nabd.health/errors/validation-failed" }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
    PageMeta:
      type: object
      properties:
        nextCursor: { type: [string, "null"] }
        limit: { type: integer }
...
    StaffPatch:
      type: object
      properties:
        roleId: { type: string, format: uuid }
        scope: { type: string, enum: [own_patients_only, all_clinic_patients] }
        fieldGrants:
          type: array
          items: { type: string, example: "no_financial_fields" }
    Staff:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        mobilePhone: { type: [string, "null"] }
        roleId: { type: string, format: uuid }
        status: { type: string, enum: [invited, active, suspended] }
        scope: { type: string, enum: [own_patients_only, all_clinic_patients] }
        emailVerified: { type: boolean }
        mobileVerified: { type: boolean, description: "Patient data is unreachable until both this and emailVerified are true (NB-041)" }
        lastSeenAt: { type: [string, "null"], format: date-time }
    RoleWrite:
      type: object
      required: [name, grants]
      properties:
        name: { type: string }
        grants:
          type: array
          description: One entry per module the role touches
          items: { $ref: "#/components/schemas/ModuleGrant" }
    Role:
      allOf:
        - $ref: "#/components/schemas/RoleWrite"
        - type: object
          properties:
            id: { type: string, format: uuid }
            builtIn: { type: boolean }
    ModuleGrant:
      type: object
      required: [module]
      properties:
        module:
          type: string
          enum: [patients, queue, clinical, billing, packages, pharmacy, reports, staff, setup, compliance]
        view: { type: boolean, default: false }
        create: { type: boolean, default: false }
        edit: { type: boolean, default: false }
        delete: { type: boolean, default: false }
        approve: { type: boolean, default: false }
        refundDiscount: { type: boolean, default: false }
        export: { type: boolean, default: false }
...
    WorkingHoursWrite:
      type: object
      required: [dayOfWeek, startTime, endTime]
      properties:
        dayOfWeek: { type: integer, minimum: 0, maximum: 6, description: "0=Sunday..6=Saturday" }
        startTime: { type: string, example: "09:00" }
        endTime: { type: string, example: "17:00" }
        slotMinutes: { type: integer, default: 15, minimum: 1 }
        maxPatients: { type: [integer, "null"], minimum: 1, description: "Session cap shared by scheduled + walk-in patients together; null = uncapped (NB-098)" }
    WorkingHours:
      allOf:
        - $ref: "#/components/schemas/WorkingHoursWrite"
        - type: object
          properties: { id: { type: string, format: uuid } }
    DoctorLeaveWrite:
      type: object
      required: [dateFrom, dateTo]
      properties:
        dateFrom: { type: string, format: date }
        dateTo: { type: string, format: date }
        reason: { type: string }
    DoctorLeave:
      allOf:
        - $ref: "#/components/schemas/DoctorLeaveWrite"
        - type: object
          properties: { id: { type: string, format: uuid } }
    AppointmentWrite:
      type: object
      required: [patientId, doctorId, startTime]
      properties:
        patientId: { type: string, format: uuid }
        doctorId: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
    Appointment:
      type: object
      properties:
        id: { type: string, format: uuid }
        patientId: { type: string, format: uuid }
        doctorId: { type: string, format: uuid }
        startTime: { type: string, format: date-time }
        endTime: { type: string, format: date-time }
        status: { type: string, enum: [scheduled, cancelled, completed, no_show] }
    CheckInRequest:
      type: object
      required: [patientId, doctorId]
      properties:
        appointmentId: { type: [string, "null"], format: uuid, description: "Omit for a walk-in" }
        patientId: { type: string, format: uuid }
        doctorId: { type: string, format: uuid }
    QueueEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        appointmentId: { type: [string, "null"], format: uuid }
        patientId: { type: string, format: uuid }
        doctorId: { type: string, format: uuid }
        queueDate: { type: string, format: date }
        tokenNumber: { type: integer }
        status: { type: string, enum: [checked_in, waiting, vitals_pending, vitals_done, in_consult, checkout_pending, completed, no_show] }
        priority: { type: boolean }
        priorityReason: { type: [string, "null"] }
...